pyproject.toml
[project]
name = "trading-system"
version = "0.1.0"
description = "Intraday scalp scanner and overnight BTST analyzer for Indian equity markets (NSE)"
requires-python = ">=3.14"
dependencies = [
    "yfinance",
    "pandas",
    "numpy",
    "scipy",
    "scikit-learn",
    "pyyaml",
    "python-dotenv",
    "peewee",
    "fastapi",
    "uvicorn",
    "requests",
    "openai",
    "supabase",
    "rich",
    "beautifulsoup4",
    "upstox-python-sdk",
    "psycopg2-binary>=2.9.11",
    "websocket-client",
    "protobuf",
]

[tool.uv]
package = false
